% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_delta}
\alias{compute_delta}
\title{Compute the toggle gap delta_i(psi) for binary misclassification
delta_i = mu(gamma + alpha'x_i) - mu(alpha'x_i)}
\usage{
compute_delta(psi, x, mu_fun)
}
\description{
Compute the toggle gap delta_i(psi) for binary misclassification
delta_i = mu(gamma + alpha'x_i) - mu(alpha'x_i)
}
\keyword{internal}
