% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_Ihat}
\alias{compute_Ihat}
\title{Compute I_hat(psi) = (1/n) sum dot_mu(tilde_eta_i) * xi_hat * xi_hat'}
\usage{
compute_Ihat(psi, z_hat, x, mu_dot_fun)
}
\description{
Compute I_hat(psi) = (1/n) sum dot_mu(tilde_eta_i) * xi_hat * xi_hat'
}
\keyword{internal}
