% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_Ihat_multi}
\alias{compute_Ihat_multi}
\title{Compute I_hat(psi) for multicategory case}
\usage{
compute_Ihat_multi(psi, z_hat, x, K, mu_dot_fun)
}
\description{
I_hat = (1/n) sum dot_mu(eta_i) * xi_hat * xi_hat'
}
\keyword{internal}
