% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_m_bin}
\alias{compute_m_bin}
\title{Compute per-observation drift m_hat(psi) for binary misclassification}
\usage{
compute_m_bin(psi, x, mu_fun, p01, p10, pi_z)
}
\description{
Returns n x p matrix where row i is m_i(psi).
m_i = ( -c1 * delta_i,  -c2 * delta_i * x_i )
}
\keyword{internal}
