% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_Mhat_bin}
\alias{compute_Mhat_bin}
\title{Compute M_hat(psi) = d m_hat / d psi' for binary misclassification}
\usage{
compute_Mhat_bin(psi, x, mu_fun, p01, p10, pi_z)
}
\description{
M_hat is the Jacobian of m_hat w.r.t. psi (p x p matrix).
Uses numerical differentiation.
}
\keyword{internal}
