% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compute_Mhat_multi}
\alias{compute_Mhat_multi}
\title{Numerical Jacobian M_hat for multicategory drift}
\usage{
compute_Mhat_multi(psi, x, K, mu_fun, Pi, pi_z)
}
\description{
Numerical Jacobian M_hat for multicategory drift
}
\keyword{internal}
