% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_bca_multi}
\alias{fit_bca_multi}
\title{BCA for multicategory misclassification}
\usage{
fit_bca_multi(psi_naive, y, z_hat, x, K, family, Pi, pi_z)
}
\description{
BCA for multicategory misclassification
}
\keyword{internal}
