% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_bcm_multi}
\alias{fit_bcm_multi}
\title{BCM for multicategory misclassification}
\usage{
fit_bcm_multi(psi_naive, y, z_hat, x, K, family, Pi, pi_z)
}
\description{
BCM for multicategory misclassification
}
\keyword{internal}
