% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_cs_bin}
\alias{fit_cs_bin}
\title{Corrected-score estimator for binary misclassification}
\usage{
fit_cs_bin(psi_init, y, z_hat, x, family, p01, p10, pi_z)
}
\description{
Solves (1/n) sum phi_i(psi) = 0, where
phi_i = xi_hat_i * (Y_i - mu(tilde_eta_i)) - m_i(psi)
}
\keyword{internal}
