% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_cs_multi}
\alias{fit_cs_multi}
\title{Corrected-score estimator for multicategory misclassification}
\usage{
fit_cs_multi(psi_init, y, z_hat, x, K, family, Pi, pi_z)
}
\description{
Corrected-score estimator for multicategory misclassification
}
\keyword{internal}
