% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_naive_bin}
\alias{fit_naive_bin}
\title{Fit the naive GLM estimator (binary misclassification)}
\usage{
fit_naive_bin(y, z_hat, x, family)
}
\description{
Fit the naive GLM estimator (binary misclassification)
}
\keyword{internal}
