% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_naive_multi}
\alias{fit_naive_multi}
\title{Fit the naive GLM estimator (multicategory misclassification)}
\usage{
fit_naive_multi(y, z_hat, x, K, family)
}
\description{
Fit the naive GLM estimator (multicategory misclassification)
}
\keyword{internal}
