% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcov_bc_multi}
\alias{vcov_bc_multi}
\title{Sandwich variance for BCA/BCM estimators (multicategory)}
\usage{
vcov_bc_multi(
  psi_bc,
  y,
  z_hat,
  x,
  K,
  family,
  Pi = NULL,
  pi_z = NULL,
  psi_naive = NULL,
  type = c("bca", "bcm"),
  corrected = FALSE
)
}
\description{
Sandwich variance for BCA/BCM estimators (multicategory)
}
\keyword{internal}
