% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcov_cs_bin}
\alias{vcov_cs_bin}
\title{Sandwich variance for corrected-score estimator (binary)}
\usage{
vcov_cs_bin(psi, y, z_hat, x, family, p01, p10, pi_z)
}
\description{
V_cs = J^{-1} S (J^{-1})' / n, where
J = (1/n) sum d phi_i / d psi',
S = (1/n) sum phi_i phi_i'.
}
\keyword{internal}
