% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcov_cs_multi}
\alias{vcov_cs_multi}
\title{Sandwich variance for corrected-score estimator (multicategory)}
\usage{
vcov_cs_multi(psi, y, z_hat, x, K, family, Pi, pi_z)
}
\description{
Sandwich variance for corrected-score estimator (multicategory)
}
\keyword{internal}
