% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcov_naive}
\alias{vcov_naive}
\title{Sandwich variance for the naive estimator}
\usage{
vcov_naive(psi, y, z_hat, x, family)
}
\description{
V_naive = A^(-1) C A^(-1) / n, where
A = (1/n) sum dot_mu(eta_i) xi_hat xi_hat',
C = (1/n) sum eps_i^2 xi_hat xi_hat'.
}
\keyword{internal}
