% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance.R
\name{vcov_naive_multi}
\alias{vcov_naive_multi}
\title{Sandwich variance for naive estimator (multicategory)}
\usage{
vcov_naive_multi(psi, y, z_hat, x, K, family)
}
\description{
Sandwich variance for naive estimator (multicategory)
}
\keyword{internal}
