% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_bca_bin}
\alias{fit_bca_bin}
\title{Additive bias correction (BCA) for binary misclassification}
\usage{
fit_bca_bin(psi_naive, y, z_hat, x, family, p01, p10, pi_z)
}
\description{
psi_bca = psi_hat - I_hat(psi_hat)^(-1) * m_hat(psi_hat)
}
\keyword{internal}
