% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimators.R
\name{fit_bcm_bin}
\alias{fit_bcm_bin}
\title{Multiplicative bias correction (BCM) for binary misclassification}
\usage{
fit_bcm_bin(psi_naive, y, z_hat, x, family, p01, p10, pi_z)
}
\description{
psi_bcm = psi_hat - (I_hat + M_hat)^(-1) * m_hat
}
\keyword{internal}
